# api/openapi.yaml
openapi: 3.0.3
info:
  title: Dredge API
  version: 0.1.0
  description: |
    WebSocket (live events): GET `/ws` with header `Authorization: Bearer <JWT>`,
    or query `?token=<JWT>` (required for browser WebSockets that cannot set headers).
    Upgrade is handled outside this OpenAPI surface; clients should connect to the same base URL as `servers`.
servers:
  - url: http://localhost:8080
paths:
...
  /settings/twitch-accounts:
    get:
      operationId: listTwitchAccounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Twitch accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TwitchAccount"
    post:
      operationId: createTwitchAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTwitchAccountRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwitchAccount"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TwitchAccount:
      type: object
      required: [id, username, account_type, created_at]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        account_type:
          type: string
          enum: [main, bot]
        created_at:
          type: string
          format: date-time
    CreateTwitchAccountRequest:
      type: object
      required: [id, username, refresh_token]
      properties:
        id:
          type: integer
          format: int64
          description: Twitch user id for this OAuth identity (same as Helix user id)
        username:
          type: string
        refresh_token:
          type: string
        account_type:
          type: string
          enum: [main, bot]
          default: main
...